+++ src/V1/ServiceBricks.ServiceBus.Azure/Model/ServiceBusTopic.cs
@@ -1,6 +1,7 @@
 ﻿using Azure.Messaging.ServiceBus;
 using Azure.Messaging.ServiceBus.Administration;
 using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
 using System.Text;
@@ -17,7 +18,7 @@ namespace ServiceBricks.ServiceBus.Azure
         protected readonly IBusinessRuleRegistry _businessRuleRegistry;
         protected readonly IServiceBusConnection _serviceBusConnection;
         protected readonly IConfiguration _configuration;
-        protected readonly IBusinessRuleService _businessRuleService;
+        protected readonly IServiceProvider _serviceProvider;
 
         protected ServiceBusProcessor _processor;
 
@@ -39,21 +40,21 @@ namespace ServiceBricks.ServiceBus.Azure
         /// <param name="businessRuleRegistry"></param>
         /// <param name="configuration"></param>
         /// <param name="serviceBusConnection"></param>
-        /// <param name="businessRuleService"></param>
+        /// <param name="serviceProvider"></param>
         public ServiceBusTopic(
             ILoggerFactory loggerFactory,
             IServiceBusQueue serviceBusQueue,
             IBusinessRuleRegistry businessRuleRegistry,
             IConfiguration configuration,
             IServiceBusConnection serviceBusConnection,
-            IBusinessRuleService businessRuleService)
+            IServiceProvider serviceProvider)
         {
             _logger = loggerFactory.CreateLogger<ServiceBusTopic>();
             _serviceBusQueue = serviceBusQueue;
             _businessRuleRegistry = businessRuleRegistry;
             _configuration = configuration;
             _serviceBusConnection = serviceBusConnection;
-            _businessRuleService = businessRuleService;
+            _serviceProvider = serviceProvider;
 
             // Set topic
             var topic = _configuration.GetValue<string>(ServiceBusAzureConstants.APPSETTINGS_TOPIC);
@@ -347,7 +348,7 @@ namespace ServiceBricks.ServiceBus.Azure
                     string messageData = args.Message.Body.ToString();
 
                     // Process the broadcast
-                    if (await ProcessBroadcast(eventName, messageData))
+                    if (await ProcessBroadcastAsync(eventName, messageData))
                         await args.CompleteMessageAsync(args.Message);
                     else
                         await args.DeadLetterMessageAsync(args.Message);
@@ -374,7 +375,7 @@ namespace ServiceBricks.ServiceBus.Azure
         /// <param name="broadcastName"></param>
         /// <param name="message"></param>
         /// <returns></returns>
-        protected virtual async Task<bool> ProcessBroadcast(string broadcastName, string message)
+        protected virtual async Task<bool> ProcessBroadcastAsync(string broadcastName, string message)
         {
             // If subscribed to, a reference will be found
             var type = _businessRuleRegistry.GetKeys().Where(x => x.FullName == broadcastName).FirstOrDefault();
@@ -383,10 +384,14 @@ namespace ServiceBricks.ServiceBus.Azure
                 // Convert to the broadcast type
                 var domainBroadcast = (IDomainBroadcast)JsonConvert.DeserializeObject(message, type);
 
-                // Execute the broadcast
-                BusinessRuleContext context = new BusinessRuleContext(domainBroadcast);
-                var resp = await _businessRuleService.ExecuteRulesAsync(context);
-                return resp.Success;
+                using (var scope = _serviceProvider.CreateScope())
+                {
+                    // Execute the broadcast
+                    BusinessRuleContext context = new BusinessRuleContext(domainBroadcast);
+                    var businessRuleService = scope.ServiceProvider.GetRequiredService<IBusinessRuleService>();
+                    var resp = await businessRuleService.ExecuteRulesAsync(context);
+                    return resp.Success;
+                }
             }
             return true;
         }